moved unnecesary local function declaration to global context
